# openapi/api-v1.yaml
openapi: 3.1.0
info:
  title: ILM Red API
  description: |
    Cloud-native API for digital knowledge management platform.

    ## Authentication

    The API supports two authentication methods:

    1. **Bearer Token (JWT)**: Include the JWT in the `Authorization` header
       ```
       Authorization: Bearer eyJhbGciOiJSUzI1NiI...
       ```

    2. **API Key**: Include your API key in the `X-API-Key` header
       ```
       X-API-Key: ilm_live_abc123...
       ```

    ## Rate Limiting

    | Tier       | Requests/min | AI Tokens/day |
    |------------|--------------|---------------|
    | Free       | 60           | 10,000        |
    | Premium    | 300          | 100,000       |
    | Enterprise | 1,000        | Unlimited     |

    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit`
    - `X-RateLimit-Remaining`
    - `X-RateLimit-Reset`

    ## Pagination

    List endpoints support pagination with `page` and `limit` query parameters.
    Response includes a `pagination` object with total count and page info.

  version: 1.0.0
  contact:
    name: ILM Red Support
    email: support@ilm-red.com
    url: https://docs.ilm-red.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.ilm-red.com/v1
    description: Production
  - url: https://api-staging.ilm-red.com/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Development
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /users/me:
    get:
      tags:
        - Users
      summary: Get current user profile
      operationId: getCurrentUser
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'

    patch:
      tags:
        - Users
      summary: Update current user profile
      operationId: updateCurrentUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
                  maxLength: 100
                username:
                  type: string
                  pattern: '^[a-z0-9_-]{3,30}$'
                bio:
                  type: string
                  maxLength: 500
                preferences:
                  $ref: '#/components/schemas/UserPreferences'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
...
  /users/{userId}:
    get:
      tags:
        - Users
      summary: Get user public profile
      operationId: getUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUserProfile'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /admin/users/{userId}:
    patch:
      tags:
        - Admin
      summary: Update user
      operationId: adminUpdateUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [active, suspended]
                roles:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: User updated
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  responses:
...
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
            timestamp:
              type: string
              format: date-time
            documentation:
              type: string
              format: uri
      example:
        error:
          code: "BOOK_NOT_FOUND"
          message: "The requested book does not exist or you don't have access"
          requestId: "req_abc123xyz789"
          timestamp: "2025-07-15T10:30:00Z"
          documentation: "https://api.ilm-red.com/docs/errors#BOOK_NOT_FOUND"
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
          format: uri
          nullable: true
        bio:
          type: string
          nullable: true
        roles:
          type: array
          items:
            type: string
        preferences:
          $ref: '#/components/schemas/UserPreferences'
        stats:
          type: object
          properties:
            booksUploaded:
              type: integer
            booksRead:
              type: integer
            clubsJoined:
              type: integer
        createdAt:
          type: string
          format: date-time
    PublicUserProfile:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
          format: uri
          nullable: true
        bio:
          type: string
          nullable: true
      example:
        id: "61e3db52-fac1-4c00-8fbf-75287ed8084a"
        username: "reader_61e3db52"
        displayName: "Alex Smith"
        avatarUrl: null
        bio: "Book enthusiast and avid reader"
    UserPreferences:
      type: object
      properties:
        theme:
          type: string
          enum: [light, dark, system]
        language:
          type: string
        timezone:
          type: string
        notifications:
          type: object
          properties:
            email:
              type: boolean
            push:
              type: boolean
...